+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Graceful no-downtime reload in cluster mode via SIGHUP
+
 ## Release 1.1.0 (2025-02-10)
 
 - Change signature for GET BACKUP endpoint